Fix package specs & install in dependency order
I thought I could specify the spec as major.minor without a wildcard
for the patch, but apparently not. This might be package specific,
so I may need another commit or two to fix this. Also, since CI
was throwing errors due to this issue already, I reverted the out
of order dependency installation.